requirements.txt
torch>=2.4.1
torchvision>=0.19.1
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.22.0
scikit-learn>=1.3.0
lpips>=0.1.4
pytorch-fid>=0.3.0
pytorch-wavelets>=1.3.0
PyWavelets>=1.6.0
thop>=0.1.1
psutil>=5.9.0
